+++ helper-scripts/src/stackManagementTool/stackManagementTool.ts
@@ -1,9 +1,9 @@
 import { echo } from "zx";
-import { deployStackTool } from "../deploy/deployStackTool.js";
+import { deployStackTool } from "./deploy/deployStackTool.js";
+import { whichDeletionTool } from "./delete/selectDeletionTool/selectDeletionTool.js";
 import { askWhichTool, welcomeMessage } from "./prompts.js";
-import { whichDeletionTool } from "../delete/selectDeletionTool/selectDeletionTool.js";
 
-export const selectTool = async (): Promise<void> => {
+export const stackManagementTool = async (): Promise<void> => {
   welcomeMessage();
 
   const { choice } = await askWhichTool();
